test folder renames
