+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/demo.routing.module.ts
@@ -1,30 +1,29 @@
 import { Routes } from '@angular/router';
 import { DemoContainerComponent } from './demo-container/demo-container.component';
-import { AsyncComponent } from './samples/async/async.component';
+import { AsyncComponent } from './samples/component-async/async.component';
 import { ComponentClassComponent } from './samples/component-class/component-class.component';
 import { ComponentEventsComponent } from './samples/component-events/component-events.component';
+import { IntegrationTestComponent } from './samples/component-integration/integration-test.component';
+import { MarblesComponent } from './samples/component-marbles/marbles.component';
 import { ComponentTestComponent } from './samples/component-test/component-test.component';
 import { ComponentWriteComponent } from './samples/component-write/component-write.component';
 import { CypressComponent } from './samples/cypress/cypress.component';
-import { DirectiveComponent } from './samples/directive/directive.component';
+import { DirectiveHostComponent } from './samples/directive/directive-host/directive-host.component';
+import { HttpTestsBsComponent } from './samples/http-tests-bs/http-tests-bs.component';
 import { HttpTestsComponent } from './samples/http-tests/http-tests.component';
-import { IntegrationTestComponent } from './samples/integration-tests/integration-test.component';
 import { UnitTestingComponent } from './samples/intro-unit-testing/unit-testing.component';
-import { MarblesComponent } from './samples/marbles/marbles.component';
 import { MaterialAsyncComponent } from './samples/material-async/material-async.component';
 import { MaterialComponent } from './samples/material/material.component';
 import { MockStoreComponent } from './samples/mock-store/mock-store.component';
 import { MockHostComponent } from './samples/mocking/mock-host/mock-host.component';
 import { SpyHostComponent } from './samples/mocking/spy-host/spy-host.component';
+import { NgrxMockSelectorsComponent } from './samples/ngrx-mock-selectors/ngrx-mock-selectors.component';
+import { NgrxMockstoreComponent } from './samples/ngrx-mockstore/ngrx-mockstore.component';
+import { NgrxReducersComponent } from './samples/ngrx-reducers/ngrx-reducers.component';
+import { NgrxSelectorsComponent } from './samples/ngrx-selectors/ngrx-selectors.component';
 import { NgrxComponent } from './samples/ngrx/ngrx.component';
 import { TestPipeComponent } from './samples/pipe/test-pipe.component';
 import { SimpleServiceComponent } from './samples/simple-service/simple-service.component';
-import { DirectiveHostComponent } from './samples/directive/directive-host/directive-host.component';
-import { NgrxMockstoreComponent } from './samples/ngrx-mockstore/ngrx-mockstore.component';
-import { NgrxMockSelectorsComponent } from './samples/ngrx-mock-selectors/ngrx-mock-selectors.component';
-import { NgrxSelectorsComponent } from './samples/ngrx-selectors/ngrx-selectors.component';
-import { NgrxReducersComponent } from './samples/ngrx-reducers/ngrx-reducers.component';
-import { HttpTestsBsComponent } from './samples/http-tests-bs/http-tests-bs.component';
 
 export const demoRoutes: Routes = [
   {
