--- demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/async/async.component.html
@@ -1,11 +1,0 @@
-<app-markdown-renderer [md]="'async'"></app-markdown-renderer>
-<!-- 
-<app-simple-auth-async></app-simple-auth-async>
-
-<app-simple-auth-done></app-simple-auth-done>
-
-<app-simple-auth-fake-async></app-simple-auth-fake-async>
-
-<app-simple-auth-when-stable></app-simple-auth-when-stable> -->
-
-<app-timer></app-timer>
