+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-async/async.component.html
@@ -1,0 +1,11 @@
+<app-markdown-renderer [md]="'component-async'"></app-markdown-renderer>
+
+<app-timer></app-timer>
+
+<app-simple-auth-async></app-simple-auth-async>
+
+<app-simple-auth-done></app-simple-auth-done>
+
+<app-simple-auth-fake-async></app-simple-auth-fake-async>
+
+<app-simple-auth-when-stable></app-simple-auth-when-stable>
