+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-async/async.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { SimpleAuthService } from './simple-auth.service';
 
 @Component({
   selector: 'app-async',
@@ -7,7 +6,7 @@ import { SimpleAuthService } from './simple-auth.service';
   styleUrls: ['./async.component.scss'],
 })
 export class AsyncComponent implements OnInit {
-  constructor() {}
+  constructor() { }
 
-  ngOnInit() {}
+  ngOnInit() { }
 }
