+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-async/simple-auth-async-when-stable/simple-auth-when-stable.component.html
