+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-async/timer/timer.component.ts
