+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-integration/food-list/food-list.component.spec.ts
@@ -1,11 +1,11 @@
-import { NO_ERRORS_SCHEMA, DebugElement } from '@angular/core';
+import { DebugElement, NO_ERRORS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { of } from 'rxjs';
-import { FoodListComponent } from './food-list.component';
-import { FoodRowComponent } from '../food-row/food-row.component';
-import { RatingPipe } from '../../pipe/rating.pipe';
 import { FoodServiceBS } from '../../food/food.service-bs';
+import { RatingPipe } from '../../pipe/rating.pipe';
+import { FoodRowComponent } from '../food-row/food-row.component';
+import { FoodListComponent } from './food-list.component';
 
 describe('Component - Integration Test', () => {
   let fs: any;
@@ -65,6 +65,6 @@ describe('Component - Integration Test', () => {
 
     fixture.detectChanges();
     const rows = de.queryAll(By.directive(FoodRowComponent));
-    expect(rows.length).toEqual(4);
+    expect(rows.length).toEqual(3);
   });
 });
