+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-integration/food-row/food-row.component.html
@@ -1,4 +1,4 @@
-<div class="hoverclick" fxLayout="row" fxLayoutGap="10px">
+<div class="foodrow" fxLayout="row" fxLayoutGap="10px">
   <div id="itemName" fxFlex="3 1 300">{{ food?.name }}:</div>
   <div id="rating" fxFlex="2 1 100">{{ food!=null ? food.rating : "" }}</div>
   <div fxFlex="1 1 60" (click)="deleteFood()" id="deleteFld">
