+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-integration/food-row/food-row.component.scss
@@ -1,0 +1,9 @@
+.foodrow {
+  cursor: pointer;
+  margin: 0.2rem;
+  border-bottom: solid 1px black;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
