+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-integration/integration-test.component.html
@@ -1,4 +1,4 @@
-<app-markdown-renderer [md]="'shallow'"></app-markdown-renderer>
+<app-markdown-renderer [md]="'component-integration'"></app-markdown-renderer>
 
 <mat-card>
   <mat-card-header>
@@ -11,6 +11,6 @@
   </mat-card-content>
 </mat-card>
 
-<app-markdown-renderer [md]="'shallow2'"></app-markdown-renderer>
+<app-markdown-renderer [md]="'component-integration-2'"></app-markdown-renderer>
 
 <app-food-list></app-food-list>
