+++ demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/component-marbles/marble-syntax.spec.ts
