--- demos/08-testing/01-jasmine/jasmine-testing/src/app/demos/samples/integration-tests/food-row/food-row.component.scss
@@ -1,15 +1,0 @@
-
-:host {
-  width: 100%;
-}
-
-.hoverclick {
-  cursor: pointer;
-  background-color:var(--color-primary);
-  color: white;
-  margin: 0.2rem;
-  padding: 0.2rem;
-  &:hover {
-    text-decoration: underline;
-  }
-}
