+++ demos/08-testing/01-jasmine/jasmine-testing/src/assets/markdown/component-async.md
@@ -1,5 +1,7 @@
-Examine the `*.spec` files in `./demo/samples/async`.
+Examine the `*.spec` files in `/component-async`.
 
 Examine `async-intro.spec.ts` and try to understand it
 
 Examine `timer.component.spec.ts`
+
+Examine the Simple Auth implementations
