+++ demos/08-testing/01-jasmine/jasmine-testing/src/assets/markdown/component-integration-2.md
