+++ demos/08-testing/01-jasmine/jasmine-testing/src/assets/markdown/component-integration.md
@@ -1,0 +1,5 @@
+Navigate to folder \component-integration
+
+Examine `integration-test.component.html`. Notice the `app-food-row` component and inspect it.
+
+Take a look at `food-row.component.spec.ts`
