+++ demos/08-testing/01-jasmine/jasmine-testing/src/assets/markdown/marbles.md
@@ -1,4 +1,4 @@
-Investigate the contet of the marbles folder
+Examine the `*.spec` files in `/component-marbles`.
 
 - marble-syntax.spec.ts
 - user-marbles.component.spec.ts
