--- demos/08-testing/01-jasmine/jasmine-testing/src/assets/markdown/shallow.md
@@ -1,5 +1,0 @@
-Navigate to folder \demos\integration-tests
-
-Investigate `integration-test.component.html`. Notice the `app-food-row` component and instpect it
-
-Take a look at `food-row.component.spec.ts`
